chore: remove unused ACPowerStateChange code
The toggling of circuit 79 is a remnant from PR #796. In PR #2119 Harry
changed circuit 79 to refer to a different element. On requesting
information Harry indicated not being aware of this code, nor of how to
test if this code is needed. Therefore I'll simply remove it and we'll
see what happens...