feat: analog BRK gauge signal source
The signals to the BRK gauge come from DC ESS powered sources. Therefore
we check for DC ESS power to be available in order to determine what
to display.

As the BRK gauge is an analog gauge, I removed the condition of it being
only displayed when there is power. There is still an issue with this:
the panel is a source of light in a C&D cockpit while it shouldn't be.